Added licensing information for worker, added comments to schema
